--- src/core/components/formContainer/index.ts
@@ -1,2 +1,0 @@
-export { FormContainer } from './FormContainer';
-export type { FormContainerProps } from './types';
