--- src/core/components/formContainer/types.ts
@@ -1,7 +1,0 @@
-import type { ReactNode } from 'react';
-
-export interface FormContainerProps {
-  label?: string;
-  children?: ReactNode;
-  errorMessage?: string;
-}
