+++ src/core/components/formGroup/FormGroup.tsx
@@ -1,17 +1,28 @@
-import cn from 'classnames';
 import styles from './FormGroup.module.css';
+import type { ReactNode } from 'react';
 
-import type { FormGroupProps } from './types';
-import { FormContainer } from '../formContainer';
+export interface FormGroupProps {
+  title: string;
+  subTitle: string;
+  label: string;
+  children: ReactNode;
+  errorMessage?: string;
+}
 
-export const FormGroup = ({ title, subTitle, label, children, errorMessage }: FormGroupProps) => {
+export const FormGroup = ({ title, subTitle, label, errorMessage = '', children }: FormGroupProps) => {
   return (
-    <fieldset className={cn(styles.formGroup)}>
-      <div className={cn(styles.titleContainer)}>
-        <div className={cn(styles.title)}>{title}</div>
-        <div className={cn(styles.subTitle)}>{subTitle}</div>
+    <fieldset className={styles.formGroup}>
+      <legend className={styles.legend}>
+        <div className={styles.title}>{title}</div>
+        <div className={styles.subTitle}>{subTitle}</div>
+      </legend>
+      <div className={styles.formWrapper}>
+        <span className={styles.label}>{label}</span>
+        <div className={styles.children}>{children}</div>
+        <div className={styles.errorMessage} role="alert">
+          {errorMessage}
+        </div>
       </div>
-      <FormContainer label={label} children={children} errorMessage={errorMessage} />
     </fieldset>
   );
 };
