+++ src/core/components/formGroup/index.ts
@@ -1,2 +1,2 @@
 export { FormGroup } from './FormGroup';
-export type { FormGroupProps } from './types';
+export type { FormGroupProps } from './FormGroup';
