--- src/core/components/formGroup/types.ts
@@ -1,6 +1,0 @@
-import type { FormContainerProps } from '../formContainer';
-
-export interface FormGroupProps extends FormContainerProps {
-  title?: string;
-  subTitle?: string;
-}
